+++ src/CodeGenerator/InjectorGenerator.php
@@ -7,16 +7,24 @@
 
 namespace Zend\Di\CodeGenerator;
 
-use Zend\Code\Generator\ClassGenerator;
-use Zend\Code\Generator\DocBlockGenerator;
-use Zend\Code\Generator\FileGenerator;
-use Zend\Code\Generator\MethodGenerator;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use SplFileObject;
+use Throwable;
 use Zend\Di\ConfigInterface;
 use Zend\Di\Definition\DefinitionInterface;
 use Zend\Di\Resolver\DependencyResolverInterface;
 
+use function array_keys;
+use function array_map;
+use function file_get_contents;
+use function implode;
+use function str_repeat;
+use function strtr;
+use function var_export;
+
 /**
- * Generator for the depenendency injector
+ * Generator for the dependency injector
  *
  * Generates a Injector class that will use a generated factory for a requested
  * type, if available. This factory will contained pre-resolved dependencies
@@ -25,6 +33,10 @@ use Zend\Di\Resolver\DependencyResolverInterface;
 class InjectorGenerator
 {
     use GeneratorTrait;
+
+    private const FACTORY_LIST_TEMPLATE = __DIR__ . '/../../templates/factory-list.template';
+    private const INJECTOR_TEMPLATE = __DIR__ . '/../../templates/injector.template';
+    private const INDENTATION_SPACES = 4;
 
     /**
      * @var ConfigInterface
@@ -43,11 +55,6 @@ class InjectorGenerator
     protected $definition;
 
     /**
-     * @var int
-     */
-    private $factoryIndex = 0;
-
-    /**
      * @var string
      */
     private $namespace;
@@ -63,62 +70,78 @@ class InjectorGenerator
     private $autoloadGenerator;
 
     /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
      * Constructs the compiler instance
      *
      * @param ConfigInterface $config The configuration to compile from
      * @param DependencyResolverInterface $resolver The resolver to utilize
-     * @param string $namespace Namespace to use for generated class; defaults
+     * @param string|null $namespace Namespace to use for generated class; defaults
      *     to Zend\Di\Generated.
+     * @param LoggerInterface|null $logger An optional logger instance to log failures
+     *     and processed classes.
      */
     public function __construct(
         ConfigInterface $config,
         DependencyResolverInterface $resolver,
-        ?string $namespace = null
+        string $namespace = null,
+        LoggerInterface $logger = null
     ) {
         $this->config = $config;
         $this->resolver = $resolver;
         $this->namespace = $namespace ? : 'Zend\Di\Generated';
         $this->factoryGenerator = new FactoryGenerator($config, $resolver, $this->namespace . '\Factory');
         $this->autoloadGenerator = new AutoloadGenerator($this->namespace);
+        $this->logger = $logger ?? new NullLogger();
     }
 
-    /**
-     * Generate injector
-     *
-     * @param array $factories
-     */
-    private function generateInjector(array $factories)
+    private function buildFromTemplate(string $templateFile, string $outputFile, array $replacements) : void
     {
-        $listFile = new FileGenerator();
-        $listFile->setFilename($this->outputDirectory . '/factories.php')
-            ->setDocBlock(new DocBlockGenerator('AUTO GENERATED FACTORY LIST'))
-            ->setBody('return ' . var_export($factories, true) . ';');
+        $template = file_get_contents($templateFile);
+        $code = strtr($template, $replacements);
+        $file = new SplFileObject($outputFile, 'w');
 
-        $class = new ClassGenerator('GeneratedInjector', $this->namespace);
-        $classFile = new FileGenerator();
-
-        $loadFactoryCode = '$this->factories = require __DIR__ . \'/factories.php\';';
-        $class->setExtendedClass('\\' . AbstractInjector::class)
-            ->addMethod('loadFactoryList', [], MethodGenerator::FLAG_PUBLIC, $loadFactoryCode);
-
-        $classFile->setFilename($this->outputDirectory . '/GeneratedInjector.php')
-            ->setDocBlock(new DocBlockGenerator('AUTO GENERATED DEPENDENCY INJECTOR'))
-            ->setNamespace($class->getNamespaceName())
-            ->setClass($class);
-
-        $listFile->write();
-        $classFile->write();
+        $file->fwrite($code);
+        $file->fflush();
     }
 
-    /**
-     * @param string $class
-     * @param array $factories
-     */
-    private function generateTypeFatory(string $class, array &$factories)
+    private function generateInjector() : void
+    {
+        $this->buildFromTemplate(
+            self::INJECTOR_TEMPLATE,
+            sprintf('%s/GeneratedInjector.php', $this->outputDirectory),
+            [
+                '%namespace%' => $this->namespace ? "namespace {$this->namespace};\n" : '',
+            ]
+        );
+    }
+
+    private function generateFactoryList(array $factories) : void
+    {
+        $indentation = sprintf("\n%s", str_repeat(' ', self::INDENTATION_SPACES));
+        $codeLines = array_map(
+            function (string $key, string $value) : string {
+                return sprintf('%s => %s,', var_export($key, true), var_export($value, true));
+            },
+            array_keys($factories),
+            $factories
+        );
+
+        $this->buildFromTemplate(self::FACTORY_LIST_TEMPLATE, sprintf('%s/factories.php', $this->outputDirectory), [
+            '%factories%' => implode($indentation, $codeLines),
+        ]);
+    }
+
+    private function generateTypeFactory(string $class, array &$factories) : void
     {
         if (isset($factories[$class])) {
             return;
         }
+
+        $this->logger->debug(sprintf('Generating factory for class "%s"', $class));
 
         try {
             $factory = $this->factoryGenerator->generate($class);
@@ -126,15 +149,16 @@ class InjectorGenerator
             if ($factory) {
                 $factories[$class] = $factory;
             }
-        } catch (\Exception $e) {
-            // TODO: logging/notifying ...
+        } catch (Throwable $e) {
+            $this->logger->error(sprintf(
+                'Could not create factory for "%s": %s',
+                $class,
+                $e->getMessage()
+            ));
         }
     }
 
-    /**
-     * @return void
-     */
-    private function generateAutoload()
+    private function generateAutoload() : void
     {
         $addFactoryPrefix = function ($value) {
             return 'Factory/' . $value;
@@ -147,13 +171,21 @@ class InjectorGenerator
     }
 
     /**
+     * Returns the namespace this generator uses
+     */
+    public function getNamespace() : string
+    {
+        return $this->namespace;
+    }
+
+    /**
      * Generate the injector
      *
      * This will generate the injector and its factories into the output directory
      *
      * @param string[] $classes
      */
-    public function generate($classes = [])
+    public function generate($classes = []) : void
     {
         $this->ensureOutputDirectory();
         $this->factoryGenerator->setOutputDirectory($this->outputDirectory . '/Factory');
@@ -161,14 +193,15 @@ class InjectorGenerator
         $factories = [];
 
         foreach ($classes as $class) {
-            $this->generateTypeFatory((string)$class, $factories);
+            $this->generateTypeFactory((string)$class, $factories);
         }
 
         foreach ($this->config->getConfiguredTypeNames() as $type) {
-            $this->generateTypeFatory($type, $factories);
+            $this->generateTypeFactory($type, $factories);
         }
 
         $this->generateAutoload();
-        $this->generateInjector($factories);
+        $this->generateInjector();
+        $this->generateFactoryList($factories);
     }
 }