+++ test/CodeGenerator/InjectorGeneratorTest.php
@@ -8,12 +8,13 @@
 namespace ZendTest\Di\CodeGenerator;
 
 use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Psr\Log\LoggerInterface;
 use Zend\Di\CodeGenerator\InjectorGenerator;
 use Zend\Di\Config;
 use Zend\Di\Definition\RuntimeDefinition;
 use Zend\Di\Resolver\DependencyResolver;
 use ZendTest\Di\TestAsset;
-use Zend\Di\InjectorInterface;
 
 /**
  * FactoryGenerator test case.
@@ -24,7 +25,7 @@ class InjectorGeneratorTest extends TestCase
 
     use GeneratorTestTrait;
 
-    public function testGenerateCreatesFiles()
+    public function testGenerateCreatesFiles() : void
     {
         $config = new Config();
         $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
@@ -41,21 +42,82 @@ class InjectorGeneratorTest extends TestCase
         $this->assertFileExists($this->dir . '/autoload.php');
     }
 
-    public function testGeneratedInjectorIsValidCode()
+    public function testGeneratedInjectorClassCode() : void
     {
         // The namespace must be unique, Since we will attempt to load the
         // generated class
-        $namespace = self::DEFAULT_NAMESPACE . uniqid();
+        $namespace = self::DEFAULT_NAMESPACE;
         $config = new Config();
         $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
         $generator = new InjectorGenerator($config, $resolver, $namespace);
-        $class = $namespace . '\\GeneratedInjector';
 
         $generator->setOutputDirectory($this->dir);
         $generator->generate([]);
 
-        $this->assertFalse(class_exists($class, false));
-        include $this->dir . '/GeneratedInjector.php';
-        $this->assertTrue(class_exists($class, false));
+        self::assertFileEquals(
+            __DIR__ . '/../_files/expected-codegen-results/injector-class.php',
+            $this->dir . '/GeneratedInjector.php'
+        );
+    }
+
+    public function testGeneratedFactoryListCode() : void
+    {
+        // The namespace must be unique, Since we will attempt to load the
+        // generated class
+        $namespace = self::DEFAULT_NAMESPACE;
+        $config = new Config();
+        $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
+        $generator = new InjectorGenerator($config, $resolver, $namespace);
+
+        $generator->setOutputDirectory($this->dir);
+        $generator->generate([
+            TestAsset\A::class,
+            TestAsset\B::class
+        ]);
+
+        self::assertFileEquals(
+            __DIR__ . '/../_files/expected-codegen-results/factories-file.php',
+            $this->dir . '/factories.php'
+        );
+    }
+
+    public function testSetCustomNamespace() : void
+    {
+        $expected = self::DEFAULT_NAMESPACE . uniqid();
+        $config = new Config();
+        $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
+        $generator = new InjectorGenerator($config, $resolver, $expected);
+
+        $this->assertEquals($expected, $generator->getNamespace());
+    }
+
+    public function testGeneratorLogsDebugForEachClass()
+    {
+        $config = new Config();
+        $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
+        $logger = $this->prophesize(LoggerInterface::class);
+
+        $generator = new InjectorGenerator($config, $resolver, null, $logger->reveal());
+        $generator->setOutputDirectory($this->dir);
+        $generator->generate([
+            TestAsset\B::class
+        ]);
+
+        $logger->debug(Argument::containingString(TestAsset\B::class))->shouldHaveBeenCalled();
+    }
+
+    public function testGeneratorLogsErrorWhenFactoryGenerationFailed()
+    {
+        $config = new Config();
+        $resolver = new DependencyResolver(new RuntimeDefinition(), $config);
+        $logger = $this->prophesize(LoggerInterface::class);
+        $generator = new InjectorGenerator($config, $resolver, null, $logger->reveal());
+
+        $generator->setOutputDirectory($this->dir);
+        $generator->generate([
+            'Bad.And.Undefined.ClassName'
+        ]);
+
+        $logger->error(Argument::containingString('Bad.And.Undefined.ClassName'))->shouldHaveBeenCalled();
     }
 }