+++ test/_files/expected-codegen-results/autoload-file.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Generated autoload file for Zend\Di
+ */
+
+namespace ZendTest\Di\Generated;
+
+require_once __DIR__ . '/Autoloader.php';
+
+return (function(): Autoloader {
+    $autoloader = new Autoloader();
+    $autoloader->register();
+
+    return $autoloader;
+})();